filter, join & leave trips

+++ src/app/other/image-upload/image-upload.component.ts
@@ -55,12 +55,16 @@ export class ImageUploadComponent {
           this.onSuccess();
           if (this._userData.getView() === 'myprofile') {
             const user: Traveller = this._userData.getUserData();
-            this.imageService.deleteImage(user.picture);
+            if (user.picture !== 'profile-default') {
+              this.imageService.deleteImage(user.picture);
+            }
             user.picture = res['imageUrl'].replace('https://travelmanagerpictures.s3.eu-north-1.amazonaws.com/', '');
             this.travellerService.updateTraveller(user);
           } else {
             const trip: Trip = this._userData.getTripData();
-            this.imageService.deleteImage(trip.picture);
+            if (trip.picture !== 'trip-default') {
+              this.imageService.deleteImage(trip.picture);
+            }
             trip.picture = res['imageUrl'].replace('https://travelmanagerpictures.s3.eu-north-1.amazonaws.com/', '');
             this.tripService.updateTrip(trip);
           }
